Squashed all of my personalisation changes
Awesome configuration for laptop

Added conky and an installation file

Added the battery widget back into awesome

Moved IP addr conky

Added slock to to the startup script

Bigger fontsize

Added more themes to the oh-my-zsh install

Changing to the base16 colour scheme

Changed dpi for my laptop to something closer to what I actually have

Preparing for an eventual update to awesomeWM 4.x

Added mkl to the LD_LIBRARY_PATH

Different startup programs

Added a keycombo to toggle the touchpad

Move tmux away from zsh

Added spacemacs config

Changed color scheme

Pulled spacemacs config from desktop setup

Updated to awesome 4.2

lowercase awesome menus

Ruby and gem configs
