+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "X/modules/base16-shell"]
+	path = X/modules/base16-shell
+	url = https://github.com/chriskempson/base16-shell.git
