+++ X/colour_schemes/seoul256
@@ -1,0 +1,23 @@
+! seoul256 (dark) theme adapted for URxvt
+! append the content of this file to your .Xresources
+URxvt*background: #3a3a3a 
+URxvt*foreground: #d0d0d0 
+URxvt*color0: #4e4e4e
+URxvt*color1: #d68787 
+URxvt*color2: #5f865f
+URxvt*color3: #d8af5f
+URxvt*color4: #85add4
+URxvt*color5: #d7afaf
+URxvt*color6: #87afaf
+URxvt*color7: #d0d0d0
+URxvt*color8: #626262
+URxvt*color9: #d75f87
+URxvt*color10: #87af87
+URxvt*color11: #ffd787 
+URxvt*color12: #add4fb
+URxvt*color13: #ffafaf
+URxvt*color14: #87d7d7
+URxvt*color15: #e4e4e4
+URxvt*cursorColor: #d0d0d0
+URxvt*cursorColor2: #3a3a3a
+URxvt*colorBD: #e4e4e4
