+++ X/colour_schemes/solarized-dark
