+++ awesome/rc.3.5.lua
@@ -1,0 +1,629 @@
+-- Standard awesome library
+local gears = require("gears")
+local awful = require("awful")
+awful.rules = require("awful.rules")
+require("awful.autofocus")
+
+-- Widget and layout library
+local wibox = require("wibox")
+
+-- Theme handling library
+local beautiful = require("beautiful")
+
+-- Notification library
+local naughty = require("naughty")
+local menubar = require("menubar")
+local string = require("string")
+
+-- {{{ Error handling
+-- Check if awesome encountered an error during startup and fell back to
+-- another config (This code will only ever execute for the fallback config)
+if awesome.startup_errors then
+    naughty.notify({ preset = naughty.config.presets.critical,
+                     title = "Oops, there were errors during startup!",
+                     text = awesome.startup_errors })
+end
+
+-- Handle runtime errors after startup
+do
+    local in_error = false
+    awesome.connect_signal("debug::error", function (err)
+        -- Make sure we don't go into an endless error loop
+        if in_error then return end
+        in_error = true
+
+        naughty.notify({ preset = naughty.config.presets.critical,
+                         title = "Oops, an error happened!",
+                         text = err })
+        in_error = false
+    end)
+end
+-- }}}
+
+-- {{{ Variable definitions
+-- Themes define colours, icons, and wallpapers
+beautiful.init("/home/glesaaen/.config/awesome/themes/redhalo/theme.lua")
+beautiful.wallpaper = "/home/glesaaen/.config/awesome/wallpapers/wild_night.jpg"
+for s = 1, screen.count() do		
+	gears.wallpaper.centered(beautiful.wallpaper, s, "#000000")		
+end
+
+-- This is used later as the default terminal and editor to run.
+terminal = "urxvt -e tmux"
+editor = os.getenv("EDITOR") or "vim"
+editor_cmd = terminal .. " -e " .. editor
+file_browser = "pcmanfm"
+
+-- Default modkey.
+-- Usually, Mod4 is the key with a logo between Control and Alt.
+-- If you do not like this or do not have such a key,
+-- I suggest you to remap Mod4 to another key using xmodmap or other tools.
+-- However, you can use another modifier like Mod1, but it may interact with others.
+modkey = "Mod4"
+
+-- Table of layouts to cover with awful.layout.inc, order matters.
+local layouts =
+{
+    awful.layout.suit.floating,
+    awful.layout.suit.tile,
+    awful.layout.suit.tile.left,
+    awful.layout.suit.tile.bottom,
+    awful.layout.suit.tile.top,
+    awful.layout.suit.fair,
+    awful.layout.suit.fair.horizontal,
+    awful.layout.suit.spiral,
+    awful.layout.suit.spiral.dwindle,
+    awful.layout.suit.max,
+    awful.layout.suit.max.fullscreen,
+    awful.layout.suit.magnifier
+}
+-- }}}
+
+-- {{{ Tags
+-- Define a tag table which hold all screen tags.
+tags = {
+	names = { "main", "pdf", "ssh" ,"chrome", "math", "music", "lookup", "tests", "GUI" },
+	layout = { layouts[4], layouts[10], layouts[10], layouts[1], layouts[1], layouts[10],
+		   layouts[4], layouts[4], layouts[1]}
+}
+for s = 1, screen.count() do
+    -- Each screen has its own tag table.
+    tags[s] = awful.tag(tags.names, s, tags.layout)
+end
+-- }}}
+
+-- {{{ Menu
+-- Create a laucher widget and a main menu
+applications = {
+	--{"Chrome", "google-chrome-stable"},
+	{"Google Chrome", "google-chrome-stable"},
+  {"Spacemacs", "emacs"},
+  {"Google Music", "gpmdp"},
+  {"Spotify", "spotify"},
+	{"Mathematica", "mathematica"},
+	{"Skype", "skype"}
+}
+
+directories = {
+	{"Home", file_browser .. " " .. ""},
+	{"Documents", file_browser .. " " .. "Documents"},
+	{"Downloads", file_browser .. " " .. "Downloads"},
+	{"Dropbox", file_browser .. " " .. "Dropbox"},
+	{"Frankfurt", file_browser .. " " .. "Dropbox/Frankfurt"},
+	{"NTNU", file_browser .. " " .. "Dropbox/NTNU"},
+	{"Litteratur", file_browser .. " " .. "Dropbox/NTNU/Master/Litteratur"}
+}
+
+
+myawesomemenu = {
+   { "manual", "xterm -e man awesome" },
+   { "edit config", editor_cmd .. " " .. awesome.conffile },
+   { "restart", awesome.restart },
+   { "quit", awesome.quit }
+}
+
+systemcmd = {
+	{"Reboot", "systemctl reboot"},
+	{"Shutdown", "systemctl poweroff"},
+  {"Suspend", "systemctl suspend"},
+  {"Hibernate", "systemctl hibernate"}
+}
+
+mymainmenu = awful.menu({ items = { 	{ "Programs", applications},
+					{ "Dirs", directories},
+					{ "awesome", myawesomemenu, beautiful.awesome_icon },
+					{ "open terminal", terminal },
+					{ "System", systemcmd}
+                                  }
+                        })
+
+mylauncher = awful.widget.launcher({ image = beautiful.awesome_icon,
+                                     menu = mymainmenu })
+
+-- Menubar configuration
+menubar.utils.terminal = terminal -- Set the terminal for applications that require it
+-- }}}
+
+--Battery widget
+batterywidget = wibox.widget.base.make_widget()
+
+batterywidget.fit = function(batterywidget, width, height)
+  local fh = assert(io.popen("acpi | cut -d, -f 1 | cut -d: -f 2"));
+  local chargestate = string.gsub(fh:read(), "%s+", "")
+  fh:close()
+
+	local size = math.min(width, height)
+
+  if chargestate == "Charging" then
+    return 2.4*size,size
+  end
+
+	return 1.8*size, size
+end
+
+batterychargeit = 0
+
+batterywidget.draw = function(batterywidget, wibox, cr, width, height)
+  local fh = assert(io.popen("acpi | cut -d, -f 2 -", "r"))    
+	local percentage = fh:read()
+	fh:close()
+
+  local fhh = assert(io.popen("acpi | cut -d, -f 1 | cut -d: -f 2"));
+  local chargestate = string.gsub(fhh:read(), "%s+", "")
+  fhh:close()
+
+	local perNum = tonumber(string.sub(percentage,1,string.find(percentage, "%%")-1))
+	if perNum > 50 then
+		cr:set_source_rgb(0,255,0)
+	elseif perNum > 20 then
+		cr:set_source_rgb(255,211,0)
+	else
+		cr:set_source_rgb(255,0,0)
+	end
+
+  local batteryString
+
+  if chargestate == "Charging" then
+    batteryString = " |+" .. string.gsub(percentage, "%s+", "") .. "| "
+  else
+    batteryString = " |" .. string.gsub(percentage, "%s+", "") .. "| "
+  end
+
+	cr:set_font_size(11)
+	cr:move_to(0, height*0.7)
+	cr:show_text(batteryString)
+end
+
+batterytimer = timer { timeout = 5 }
+batterytimer:connect_signal("timeout", function()
+	batterywidget:emit_signal("widget::updated")
+end)
+
+batterytimer:start()
+
+--Keyboard map indicator and changer
+kbdcfg = {}
+kbdcfg.cmd = "setxkbmap"
+kbdcfg.options = '-option "ctrl:nocaps"'
+kbdcfg.layout = {"no","jp"}
+kbdcfg.layoutName = {"no","jp2"}
+kbdcfg.current = 1
+
+local layoutWidget = wibox.widget.base.make_widget()
+
+layoutWidget.fit = function(layoutWidget, width, height)
+	local size = math.min(width,height)
+	return 1.6*size, size
+end
+
+layoutWidget.draw = function(layoutWidget, wibox, cr, width, height)
+	local lang = " {" .. kbdcfg.layout[kbdcfg.current] .. "} "
+	cr:set_font_size(11)
+	cr:set_source_rgb(63,0,255)
+	cr:move_to(0, height*0.7)
+	cr:show_text(lang)
+end
+
+kbdcfg.switch = function()
+	kbdcfg.current = kbdcfg.current % #(kbdcfg.layout) + 1
+	os.execute( kbdcfg.cmd .. " " .. kbdcfg.layoutName[kbdcfg.current] .. " " .. kbdcfg.options )
+	layoutWidget:emit_signal("widget::updated")
+end
+
+--Print screen function
+screenshotloc = "/home/glesaaen/Pictures/Screenshots/" 
+
+printscreen = function()
+	local date = os.date("%d-%m-%y_%H.%M")
+	local testfile = screenshotloc..date..".png"
+	prtscr_c = 1
+	while (awful.util.file_readable(testfile) and (prtscr_c < 100)) do
+		testfile = screenshotloc..date.."_"..prtscr_c..".png"
+		prtscr_c = prtscr_c + 1
+	end
+	os.execute( "import -window root "..testfile )
+end
+
+--Touchpad toggle function
+toggle_touchpad = function()
+  local handle = io.popen("synclient -l | grep TouchpadOff | cut -d '=' -f 2 | tr -d '[:space:]'")
+  local result = handle:read("*a")
+  handle:close()
+
+  if (result == "0") then
+    os.execute("synclient TouchpadOff=1")
+  else
+    os.execute("synclient TouchpadOff=0")
+  end
+end
+
+-- {{{ Wibox
+-- Create a textclock widget
+mytextclock = awful.widget.textclock()
+
+-- Create a wibox for each screen and add it
+mywibox = {}
+mypromptbox = {}
+mylayoutbox = {}
+mytaglist = {}
+mytaglist.buttons = awful.util.table.join(
+                    awful.button({ }, 1, awful.tag.viewonly),
+                    awful.button({ modkey }, 1, awful.client.movetotag),
+                    awful.button({ }, 3, awful.tag.viewtoggle),
+                    awful.button({ modkey }, 3, awful.client.toggletag),
+                    awful.button({ }, 4, function(t) awful.tag.viewnext(awful.tag.getscreen(t)) end),
+                    awful.button({ }, 5, function(t) awful.tag.viewprev(awful.tag.getscreen(t)) end)
+                    )
+mytasklist = {}
+mytasklist.buttons = awful.util.table.join(
+                     awful.button({ }, 1, function (c)
+                                              if c == client.focus then
+                                                  c.minimized = true
+                                              else
+                                                  -- Without this, the following
+                                                  -- :isvisible() makes no sense
+                                                  c.minimized = false
+                                                  if not c:isvisible() then
+                                                      awful.tag.viewonly(c:tags()[1])
+                                                  end
+                                                  -- This will also un-minimize
+                                                  -- the client, if needed
+                                                  client.focus = c
+                                                  c:raise()
+                                              end
+                                          end),
+                     awful.button({ }, 3, function ()
+                                              if instance then
+                                                  instance:hide()
+                                                  instance = nil
+                                              else
+                                                  instance = awful.menu.clients({ width=250 })
+                                              end
+                                          end),
+                     awful.button({ }, 4, function ()
+                                              awful.client.focus.byidx(1)
+                                              if client.focus then client.focus:raise() end
+                                          end),
+                     awful.button({ }, 5, function ()
+                                              awful.client.focus.byidx(-1)
+                                              if client.focus then client.focus:raise() end
+                                          end))
+
+for s = 1, screen.count() do
+    -- Create a promptbox for each screen
+    mypromptbox[s] = awful.widget.prompt()
+    -- Create an imagebox widget which will contains an icon indicating which layout we're using.
+    -- We need one layoutbox per screen.
+    mylayoutbox[s] = awful.widget.layoutbox(s)
+    mylayoutbox[s]:buttons(awful.util.table.join(
+                           awful.button({ }, 1, function () awful.layout.inc(layouts, 1) end),
+                           awful.button({ }, 3, function () awful.layout.inc(layouts, -1) end),
+                           awful.button({ }, 4, function () awful.layout.inc(layouts, 1) end),
+                           awful.button({ }, 5, function () awful.layout.inc(layouts, -1) end)))
+    -- Create a taglist widget
+    mytaglist[s] = awful.widget.taglist(s, awful.widget.taglist.filter.all, mytaglist.buttons)
+
+    -- Create a tasklist widget
+    mytasklist[s] = awful.widget.tasklist(s, awful.widget.tasklist.filter.currenttags, mytasklist.buttons)
+
+    -- Create the wibox
+    mywibox[s] = awful.wibox({ position = "top", screen = s })
+
+    -- Widgets that are aligned to the left
+    local left_layout = wibox.layout.fixed.horizontal()
+    left_layout:add(mylauncher)
+    left_layout:add(mytaglist[s])
+    left_layout:add(mypromptbox[s])
+
+    -- Widgets that are aligned to the right
+    local right_layout = wibox.layout.fixed.horizontal()
+    if s == 1 then right_layout:add(wibox.widget.systray()) end
+    right_layout:add(batterywidget)
+    right_layout:add(layoutWidget)
+    right_layout:add(mytextclock)
+    right_layout:add(mylayoutbox[s])
+
+    -- Now bring it all together (with the tasklist in the middle)
+    local layout = wibox.layout.align.horizontal()
+    layout:set_left(left_layout)
+    layout:set_middle(mytasklist[s])
+    layout:set_right(right_layout)
+
+    mywibox[s]:set_widget(layout)
+end
+-- }}}
+
+-- {{{ Mouse bindings
+root.buttons(awful.util.table.join(
+    awful.button({ }, 3, function () mymainmenu:toggle() end),
+    awful.button({ }, 4, awful.tag.viewnext),
+    awful.button({ }, 5, awful.tag.viewprev)
+))
+-- }}}
+
+-- {{{ Key bindings
+globalkeys = awful.util.table.join(
+  awful.key({ modkey,           }, "Left",   awful.tag.viewprev       ),
+  awful.key({ modkey,           }, "Right",  awful.tag.viewnext       ),
+  awful.key({ modkey,           }, "Escape", awful.tag.history.restore),
+
+  awful.key({ modkey,           }, "j",
+    function ()
+      awful.client.focus.byidx( 1)
+      if client.focus then client.focus:raise() end
+    end),
+  awful.key({ modkey,           }, "k",
+    function ()
+      awful.client.focus.byidx(-1)
+      if client.focus then client.focus:raise() end
+    end),
+  awful.key({ modkey,           }, "w", function () mymainmenu:show() end),
+
+  -- Keyboard layout change
+	awful.key({ "Mod1" },"Shift_L", function () kbdcfg.switch() end),
+	
+  -- Print screen
+  awful.key({ }, "Print", function () printscreen() end),
+
+  -- Lock screen
+	awful.key({ "Mod1", "Control" }, "l", function () awful.util.spawn("slock") end),
+
+  -- Toggle touchpad
+  awful.key({ modkey, "Control"   }, "t", toggle_touchpad),
+
+    -- Layout manipulation
+    awful.key({ modkey, "Shift"   }, "j", function () awful.client.swap.byidx(  1)    end),
+    awful.key({ modkey, "Shift"   }, "k", function () awful.client.swap.byidx( -1)    end),
+    awful.key({ modkey, "Control" }, "j", function () awful.screen.focus_relative( 1) end),
+    awful.key({ modkey, "Control" }, "k", function () awful.screen.focus_relative(-1) end),
+    awful.key({ modkey,           }, "u", awful.client.urgent.jumpto),
+    awful.key({ modkey,           }, "Tab",
+        function ()
+            awful.client.focus.history.previous()
+            if client.focus then
+                client.focus:raise()
+            end
+        end),
+
+    -- Standard program
+    awful.key({ modkey,           }, "Return", function () awful.util.spawn(terminal) end),
+    awful.key({ modkey, "Control" }, "r", awesome.restart),
+    awful.key({ modkey, "Shift"   }, "q", awesome.quit),
+
+    awful.key({ modkey,           }, "l",     function () awful.tag.incmwfact( 0.05)    end),
+    awful.key({ modkey,           }, "h",     function () awful.tag.incmwfact(-0.05)    end),
+    awful.key({ modkey, "Shift"   }, "h",     function () awful.tag.incnmaster( 1)      end),
+    awful.key({ modkey, "Shift"   }, "l",     function () awful.tag.incnmaster(-1)      end),
+    awful.key({ modkey, "Control" }, "h",     function () awful.tag.incncol( 1)         end),
+    awful.key({ modkey, "Control" }, "l",     function () awful.tag.incncol(-1)         end),
+    awful.key({ modkey,           }, "space", function () awful.layout.inc(layouts,  1) end),
+    awful.key({ modkey, "Shift"   }, "space", function () awful.layout.inc(layouts, -1) end),
+
+    awful.key({ modkey, "Control" }, "n", awful.client.restore),
+
+    -- Prompt
+    awful.key({ modkey },            "r",     function () mypromptbox[mouse.screen]:run() end),
+
+    awful.key({ modkey }, "x",
+              function ()
+                  awful.prompt.run({ prompt = "Run Lua code: " },
+                  mypromptbox[mouse.screen].widget,
+                  awful.util.eval, nil,
+                  awful.util.getdir("cache") .. "/history_eval")
+              end),
+    -- Menubar
+    awful.key({ modkey }, "p", function() menubar.show() end)
+)
+
+clientkeys = awful.util.table.join(
+    awful.key({ modkey,           }, "f",      function (c) c.fullscreen = not c.fullscreen  end),
+    awful.key({ modkey, "Shift"   }, "c",      function (c) c:kill()                         end),
+    awful.key({ modkey, "Control" }, "space",  awful.client.floating.toggle                     ),
+    awful.key({ modkey, "Control" }, "Return", function (c) c:swap(awful.client.getmaster()) end),
+    awful.key({ modkey,           }, "o",      awful.client.movetoscreen                        ),
+    awful.key({ modkey,           }, "t",      function (c) c.ontop = not c.ontop            end),
+    awful.key({ modkey,           }, "n",
+        function (c)
+            -- The client currently has the input focus, so it cannot be
+            -- minimized, since minimized clients can't have the focus.
+            c.minimized = true
+        end),
+    awful.key({ modkey,           }, "m",
+        function (c)
+            c.maximized_horizontal = not c.maximized_horizontal
+            c.maximized_vertical   = not c.maximized_vertical
+        end)
+)
+
+-- Compute the maximum number of digit we need, limited to 9
+keynumber = 0
+for s = 1, screen.count() do
+   keynumber = math.min(9, math.max(#tags[s], keynumber))
+end
+
+-- Bind all key numbers to tags.
+-- Be careful: we use keycodes to make it works on any keyboard layout.
+-- This should map on the top row of your keyboard, usually 1 to 9.
+for i = 1, keynumber do
+    globalkeys = awful.util.table.join(globalkeys,
+        awful.key({ modkey }, "#" .. i + 9,
+                  function ()
+                        local screen = mouse.screen
+                        if tags[screen][i] then
+                            awful.tag.viewonly(tags[screen][i])
+                        end
+                  end),
+        awful.key({ modkey, "Control" }, "#" .. i + 9,
+                  function ()
+                      local screen = mouse.screen
+                      if tags[screen][i] then
+                          awful.tag.viewtoggle(tags[screen][i])
+                      end
+                  end),
+        awful.key({ modkey, "Shift" }, "#" .. i + 9,
+                  function ()
+                      if client.focus and tags[client.focus.screen][i] then
+                          awful.client.movetotag(tags[client.focus.screen][i])
+                      end
+                  end),
+        awful.key({ modkey, "Control", "Shift" }, "#" .. i + 9,
+                  function ()
+                      if client.focus and tags[client.focus.screen][i] then
+                          awful.client.toggletag(tags[client.focus.screen][i])
+                      end
+                  end))
+end
+
+clientbuttons = awful.util.table.join(
+    awful.button({ }, 1, function (c) client.focus = c; c:raise() end),
+    awful.button({ modkey }, 1, awful.mouse.client.move),
+    awful.button({ modkey }, 3, awful.mouse.client.resize))
+
+-- Set keys
+root.keys(globalkeys)
+-- }}}
+
+-- {{{ Rules
+awful.rules.rules = {
+    -- All clients will match this rule.
+    { rule = { },
+      properties = { border_width = 0,
+                     border_color = beautiful.border_normal,
+                     size_hints_honor = false,
+                     focus = awful.client.focus.filter,
+                     keys = clientkeys,
+                     buttons = clientbuttons } },
+    { rule = { class = "MPlayer" },
+      properties = { floating = true } },
+    { rule = { class = "pinentry" },
+      properties = { floating = true } },
+    { rule = { class = "Gimp" },
+      properties = { floating = true } },
+    { rule = { class = "Steam" },
+      properties = { floating = true,
+      		     tag = tags[1][9] } },
+    { rule = { class = "Google-chrome" },
+      properties = { tag = tags[1][4] } },
+    { rule = { class = "XMathematica" },
+      properties = { tag = tags[1][5] } },
+    { rule = { class = "Evince" },
+      properties = { tag = tags[1][2] } },
+    { rule = { class = "Zathura" },
+      properties = { tag = tags[1][2] } },
+    { rule = { class = "Adl" },
+      properties = { tag = tags[1][9] } },
+    -- Set Firefox to always map on tags number 2 of screen 1.
+    -- { rule = { class = "Firefox" },
+    --   properties = { tag = tags[1][2] } },
+}
+-- }}}
+
+-- {{{ Signals
+-- Signal function to execute when a new client appears.
+client.connect_signal("manage", function (c, startup)
+    -- Enable sloppy focus
+    c:connect_signal("mouse::enter", function(c)
+        if awful.layout.get(c.screen) ~= awful.layout.suit.magnifier
+            and awful.client.focus.filter(c) then
+            client.focus = c
+        end
+    end)
+
+    if not startup then
+        -- Set the windows at the slave,
+        -- i.e. put it at the end of others instead of setting it master.
+        -- awful.client.setslave(c)
+
+        -- Put windows in a smart way, only if they does not set an initial position.
+        if not c.size_hints.user_position and not c.size_hints.program_position then
+            awful.placement.no_overlap(c)
+            awful.placement.no_offscreen(c)
+        end
+    end
+
+    local titlebars_enabled = false
+    if titlebars_enabled and (c.type == "normal" or c.type == "dialog") then
+        -- Widgets that are aligned to the left
+        local left_layout = wibox.layout.fixed.horizontal()
+        left_layout:add(awful.titlebar.widget.iconwidget(c))
+
+        -- Widgets that are aligned to the right
+        local right_layout = wibox.layout.fixed.horizontal()
+        right_layout:add(awful.titlebar.widget.floatingbutton(c))
+        right_layout:add(awful.titlebar.widget.maximizedbutton(c))
+        right_layout:add(awful.titlebar.widget.stickybutton(c))
+        right_layout:add(awful.titlebar.widget.ontopbutton(c))
+        right_layout:add(awful.titlebar.widget.closebutton(c))
+
+        -- The title goes in the middle
+        local title = awful.titlebar.widget.titlewidget(c)
+        title:buttons(awful.util.table.join(
+                awful.button({ }, 1, function()
+                    client.focus = c
+                    c:raise()
+                    awful.mouse.client.move(c)
+                end),
+                awful.button({ }, 3, function()
+                    client.focus = c
+                    c:raise()
+                    awful.mouse.client.resize(c)
+                end)
+                ))
+
+        -- Now bring it all together
+        local layout = wibox.layout.align.horizontal()
+        layout:set_left(left_layout)
+        layout:set_right(right_layout)
+        layout:set_middle(title)
+
+        awful.titlebar(c):set_widget(layout)
+    end
+end)
+
+client.connect_signal("focus", function(c) c.border_color = beautiful.border_focus end)
+client.connect_signal("unfocus", function(c) c.border_color = beautiful.border_normal end)
+
+function run_once(prg,arg_string,pname,screen)
+	if not prg then
+		do return nil end
+	end
+
+	if not pname then
+		pname = prg
+	end
+
+	if not arg_string then 
+		awful.util.spawn_with_shell("pgrep -f -u $USER -x '" .. pname .. "' || (" .. prg .. ")",screen)
+	else
+		awful.util.spawn_with_shell("pgrep -f -u $USER -x '" .. pname .. " ".. arg_string .."' || (" .. prg .. " " .. arg_string .. ")",screen)
+	end
+end
+
+run_once("unclutter");
+run_once("xautolock", "-time 45 -locker slock")
+run_once("setxkbmap", 'no -option "ctrl:nocaps"')
+run_once("xflux", "-l 51.621440 -g -3.943646 -k 2000")
+run_once("compton")
+--run_once("conky")
+
+-- }}}
