+++ conky/conky.conf
@@ -1,0 +1,88 @@
+conky.config = {
+	background = true,
+	use_xft = true,
+  font = 'Roboto:size=8',
+	update_interval = 1.0,
+	total_run_times = 0,
+	own_window = true,
+	own_window_transparent = true,
+	own_window_type = 'desktop',
+	own_window_argb_visual = true,
+  own_window_hints = 'undecorated,below,skip_taskbar,sticky',
+	double_buffer = true,
+	draw_shades = false,
+	draw_outline = false,
+	draw_borders = false,
+	draw_graph_borders = false,
+	border_width = 0,
+  border_inner_margin = 40,
+	default_color = '0B486B',
+	alignment = 'top_right',
+  minimum_width = 224,
+	minimum_height = 1000,
+	maximum_width = 1000,
+	gap_x = 40,
+	gap_y = 60,
+	no_buffers = false,
+	uppercase = false,
+	cpu_avg_samples = 2,
+	net_avg_samples = 2,
+	short_units = true,
+	text_buffer_size = 2048,
+	use_spacer = 'none',
+	override_utf8_locale = true,
+}
+
+
+conky.text = [[
+
+##Time & Day
+${voffset 5}$alignc${font LG Weather_Z:size=50}${time %H:%M}
+${voffset -5}$alignc${font Aliquam:weight=Bold:size=20}${time %A, %d }
+${voffset 0}$alignc${font Aliquam:weight=Bold:size=20}${time %B }
+$hr
+# CPU Information
+${voffset -5}$alignc${font}Intel Core i7-6500U @ ${freq_g} GHz
+${voffset 8}1${offset 10}${color #0B486B}${cpubar cpu1 8,75}$color ${offset 20}2${offset 10}${color #0B486B}${cpubar cpu2 8,75}$color
+${voffset 4}3${offset 10}${color #0B486B}${cpubar cpu3 8,75}$color ${offset 20}4${offset 10}${color #0B486B}${cpubar cpu4 8,75}$color
+
+$hr
+
+# Ram, swap and storage
+ram${offset 17}${membar 8,100}${offset 10}$memfree free
+${voffset 4}swap${offset 10}${swapbar 8,100}${offset 10}$swapfree free
+${voffset 12}/${offset 37}${fs_bar 8,100 /}${offset 14}${fs_free /} free
+
+$hr
+
+# Battery
+battery ${offset 10}[${execi 60 cat /sys/class/power_supply/BAT0/status | tr '[:upper:]' '[:lower:]'} @ $battery_percent %]
+${voffset 4}${battery_bar 8,214}
+
+$hr
+
+# Network
+ssid: ${wireless_essid wlp1s0}
+${voffset 4}local ip:${goto 120}${addr wlp1s0}
+${voffset 4}external ip: ${goto 120}${execi 3600 curl -s http://icanhazip.com}
+${voffset 8}${downspeedgraph wlp1s0 12,105 0B486B 0B6B2E} ${offset 10}${upspeedgraph wlp1s0 12,105 0B486B 6B2E0B}
+${voffset 4}dl: ${downspeed wlp1s0}${goto 157} ul: ${upspeed wlp1s0}
+
+$hr
+
+#Processes
+$alignc$processes processes - $running_processes running
+
+by cpu usage ${alignr}mem    cpu%
+${color #6B2E0B}${top name 1}  $alignr${top mem_res 1}    ${top cpu 1}$color
+${top name 2}  $alignr${top mem_res 2}    ${top cpu 2}
+${top name 3}  $alignr${top mem_res 3}    ${top cpu 3}
+${top name 4}  $alignr${top mem_res 4}    ${top cpu 4}
+
+by mem usage ${alignr}mem    cpu%
+${color #6B2E0B}${top_mem name 1}  $alignr${top_mem mem_res 1}    ${top_mem cpu 1}$color
+${top_mem name 2}  $alignr${top_mem mem_res 2}    ${top_mem cpu 2}
+${top_mem name 3}  $alignr${top_mem mem_res 3}    ${top_mem cpu 3}
+${top_mem name 4}  $alignr${top_mem mem_res 4}    ${top_mem cpu 4}
+
+]]
