+++ ruby/_path.zsh
@@ -1,0 +1,1 @@
+export PATH="${HOME}/.gem/ruby/2.5.0/bin:${PATH}"
