+++ ruby/ruby.zsh
@@ -1,0 +1,2 @@
+# Set the gem location
+export GEM_HOME=${HOME}/.gem
