+++ system/env.zsh
@@ -1,5 +1,8 @@
 export EDITOR=vim
 export VIDUAL=vim
-export CC=clang
+export CC=gcc
 export CXX=clang++
+export GCC=gcc
+export MPI_HOME=/usr
 export XDG_CONFIG_HOME=$HOME/.config
+export LD_LIBRARY_PATH=/opt/intel/mkl/lib/intel64_lin
