+++ texmf/latexmkrc.symlink
@@ -1,0 +1,1 @@
+$pdflatex = 'xelatex --shell-escape %O %S'
