+++ tmux/tmux.conf.symlink
@@ -3,18 +3,20 @@ set -g default-terminal "xterm-256color"
 set -g automatic-rename off
 set -sq escape-time 0
 
-#new-session -n code zsh
-#new-window -n test zsh
-#new-window -n lookup zsh
-#new-window -t 9 -n htop htop
-
 set -g status-right ""
 set -g status-left ""
-set-option -g status-bg colour8
-set-option -g status-fg colour4
 set-window-option -g window-status-current-format "[#{window_index}:#W]"
-set-window-option -g window-status-current-fg green
 set-window-option -g xterm-keys on
+
+# Use these when I have the seoul256 colourscheme
+set-window-option -g window-status-current-fg colour3
+set-option -g status-bg colour0
+set-option -g status-fg colour1
+
+# Use these when using base16-dark
+# set-window-option -g window-status-current-fg green
+# set-option -g status-bg colour18
+# set-option -g status-fg colour5
 
 # Set the prefix to ^A.
 unbind C-b
