+++ zsh/update_colors.zsh
@@ -1,0 +1,1 @@
+#$DOTFILES/X/modules/base16-shell/scripts/base16-default-dark.sh
